requirements.txt
python-telegram-bot>=21.0
openai>=1.0
notion-client>=2.0
python-dotenv>=1.0
PyYAML>=6.0
requests>=2.28
firebase-admin>=6.0
